src/api_admin/requirements.txt
pyspark==3.4.2
python-binance
kafka-python
python-dotenv
requests
findspark
